Add API to upload pictures with corrections for exercises
A user can send a request to upload a picture with the correction of an
exercise. The picture is then encoded as a PNG file to make sure all
pictures are in the same format and same encoding. The picture is saved
on a directory with the SHA256 digest as the filename. The filename is
added to the DB for that exercise and is sent to the client when it
fetches the list of exercises in a unit.